Fix glitching low-ceiling flats (#22)
When leaving an area with a low ceiling, it was possible to glitch that
low ceiling so it rendered outside of where it physically lies. That was
because most of that flat was behind the player and the getHeight()
function overflow.

Fixed by computing radial difference between camera direction and
direction to the center of the tile. If the cos(alfa) is negative, then
it is behind the player and it doesn't have to be rendered.

Oh and I had to bring back DemoFileHandler, but only for debug builds, I
am not going to figure out demos and networking.